src/collectors/YouTubeCollector.ts
import { BaseCollector } from './BaseCollector';
import { GameDataEntry } from '../types/GameData';
import { logger } from '../utils/logger';

export class YouTubeCollector extends BaseCollector {
  constructor() {
    super('YOUTUBE' as any); // TODO: Add YOUTUBE to ContentType enum when implementing
  }

  async collect(gameName: string): Promise<GameDataEntry[]> {
    await this.rateLimit();
    
    // This is a placeholder implementation
    // The actual implementation will be driven by the tests
    throw new Error('Not implemented');
  }

  private async getTranscript(videoId: string): Promise<string> {
    throw new Error('Not implemented');
  }

  private isGamingChannel(channelTitle: string): boolean {
    throw new Error('Not implemented');
  }

  private isValidDuration(duration: string): boolean {
    throw new Error('Not implemented');
  }
}
